Move SafeDatacubeDumper to a shared position
